SWDEV-428379 - Set LD_LIBRARY_PATH for executing the binaries from build directory
The binaries rocfft_aot_helper and stockham_aot are getting executed from build directory during build time. The hardcoded RUNPATH in this binaries are helping to find the required libraries. Hardcoded RUNAPTH will be removed and will be having relative RUNPATH with respect to install directory.
LD_LIBRARY_PATH need to set during build time for a successfull exectuion of the binaries